fix(models): correctness audit + unit tests for Wave 3d det/seg models
Bug fixes uncovered during code review:

1. **RPN anchor ordering** (Faster R-CNN, Mask R-CNN, _detection.py RPN) —
   Conv2d output (B, A, H, W) was being flattened to (B, A*H*W) which is
   anchor-major order, while AnchorGenerator returns (G*A, 4) in spatial-
   major order (cell.row × cell.col × anchor). The mismatch meant logit[i]
   and anchors[i] referred to different anchors. Fixed by permuting to
   spatial-major before flattening: (B, A, H, W) → (B, H, W, A) → (B, H*W*A)
   and (B, 4A, H, W) → reshape(B, A, 4, H, W) → permute(0, 3, 4, 1, 2)
   → reshape(B, H*W*A, 4). Same fix applied in Mask R-CNN _rpn_loss().

2. **YOLOv3 channel mismatch in detection head** — _Darknet53 returns
   p3_raw at 128ch and p4_raw at 256ch (residual blocks operate at input
   channels, before the next stride-2 conv). The detection head was built
   for the standard 256/512/1024 channel layout, causing a ShapeMismatch.
   Fixed to use the actual backbone widths: concat(256+256) for P4 and
   concat(128+128) for P3.

3. **clamp() kwarg-only API mismatch** — Lucid's clamp() requires both min
   and max as positional args; .clamp(min=0.0) raises TypeError. Converted
   all single-kwarg clamps in _detection.py to positional .clamp(low, high).

4. **decode_boxes() bbox_xform_clip** — Same clamp issue; used
   .clamp(-1e9, bbox_xform_clip).

5. **Index tensor dtype** — lucid.tensor([int_list]) creates float32, but
   index_select requires int32/int64. Added .long() to all index-tensor
   construction sites in _detection.py and the four R-CNN model files.

6. **EfficientDet BiFPN transient MaxPool** — _BiFPNLayer.forward was
   creating a new nn.MaxPool2d every call. Moved to a registered
   self.down module in __init__.

New tests (lucid/test/unit/models/):

- test_models_detection.py — 54 tests covering R-CNN, Fast/Faster/Mask
  R-CNN, DETR, EfficientDet-D0, YOLO v1/v2/v3/v4. Each model verifies
  factory type, ObjectDetectionOutput shape/dtype, deterministic
  self-consistency, and no-loss-without-targets contract. Also a
  registry presence smoke test.

- test_models_segmentation.py — 24 tests for FCN, UNet variants,
  AttentionUNet, MaskFormer, Mask2Former. Verifies
  SemanticSegmentationOutput shape matches input H×W, factory types,
  and self-consistency.

All 78 tests pass; mypy --strict zero errors.

+++ lucid/models/vision/fast_rcnn/_model.py
@@ -336,7 +336,7 @@ class FastRCNNForObjectDetection(PretrainedModel):
         if not valid_mask:
             cls_loss: Tensor = lucid.zeros((1,))
         else:
-            valid_t = lucid.tensor(valid_mask)
+            valid_t = lucid.tensor(valid_mask).long()
             cls_loss = F.cross_entropy(
                 all_logits[valid_t],
                 cls_labels[valid_t],
@@ -506,7 +506,7 @@ class FastRCNNForObjectDetection(PretrainedModel):
                 if not mask:
                     continue
 
-                mask_t = lucid.tensor(mask)
+                mask_t = lucid.tensor(mask).long()
                 sc_c = cls_scores[mask_t]
                 bx_c = bx_i[mask_t]
